pyproject.toml
[tool.poetry]
name = "poc_django_rest"
version = "0.1.0"
description = ""
authors = ["AndreyMishuk <andreygim.am@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.2"
djangorestframework = "^3.11.0"
markdown = "^3.1.1"
django-filter = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
